+++ selfdrive/controls/lib/pathplanner.py
@@ -1,26 +1,29 @@
-import selfdrive.messaging as messaging
+import math
 import numpy as np
+
+from common.numpy_fast import interp
+import selfdrive.messaging as messaging
 X_PATH = np.arange(0.0, 50.0)
 
 def model_polyfit(points):
   return np.polyfit(X_PATH, map(float, points), 3)
 
 # lane width http://safety.fhwa.dot.gov/geometric/pubs/mitigationstrategies/chapter3/3_lanewidth.cfm
-_LANE_WIDTH_V = np.asarray([3., 3.8])
+_LANE_WIDTH_V = [3., 3.8]
 
 # break points of speed
-_LANE_WIDTH_BP = np.asarray([0., 31.])
+_LANE_WIDTH_BP = [0., 31.]
 
 def calc_desired_path(l_poly, r_poly, p_poly, l_prob, r_prob, p_prob, speed):
   #*** this function computes the poly for the center of the lane, averaging left and right polys
-  lane_width = np.interp(speed, _LANE_WIDTH_BP, _LANE_WIDTH_V)
+  lane_width = interp(speed, _LANE_WIDTH_BP, _LANE_WIDTH_V)
 
   # lanes in US are ~3.6m wide
   half_lane_poly = np.array([0., 0., 0., lane_width / 2.])
   if l_prob + r_prob > 0.01:
     c_poly = ((l_poly - half_lane_poly) * l_prob +
               (r_poly + half_lane_poly) * r_prob) / (l_prob + r_prob)
-    c_prob = np.sqrt((l_prob**2 + r_prob**2) / 2.)
+    c_prob = math.sqrt((l_prob**2 + r_prob**2) / 2.)
   else:
     c_poly = np.zeros(4)
     c_prob = 0.